requirements/common.txt
Django==5.0.4
mod-wsgi==5.0.0
djangorestframework==3.15.1
